Moved method from fracs.meshing to utils.
The method created circular dependencies.
